refactor: [HCM Demographics] handle dupes and change effective_at sourcing (#447)
## What does this do?

We recently requested that we receive the HCM demographics file twice daily instead of once. The reason for this is nuanced, but it's to ensure we have a record as early as possible in the day with an effective date of the current day.

There are a few things to consider here:
- The effective_at timestamp they include within the file. This timestamp sometimes shows as for prior day and sometimes as current day, depending on whether it's the morning file or evening file. This is confusing.
- The timestamp in the file name is UTC.

Ultimately this refactor removes our usage of the effective_at included within the data and instead uses the timestamp from the file name.

Additionally, we needed to handle the duplicate records which will now exist as a result of loading two files per date.

## Why do we need this?

Mainly due to duplicate records but that came up as a result of trying to improve the effective_at tracking on the records in the dataset.